# Improve shadow performance (particularly, CPU performance)