--- a/lib/ivis_opengl/piedef.h
+++ b/lib/ivis_opengl/piedef.h
@@ -69,7 +69,7 @@ void pie_setShadows(bool drawShadows);
 void pie_InitLighting();
 void pie_Lighting0(LIGHTING_TYPE entry, const float value[4]);
 
-void pie_RemainingPasses();
+void pie_RemainingPasses(UDWORD currentGameFrame);
 
 void pie_SetUp();
 void pie_CleanUp();