@past-due 2018-03-12
@perim: I agree that an intelligent shadow mapping algorithm is the logical next step. This was just low-hanging fruit.